Decompose tmux placeholder suppression helpers
