+++ ARCHITECTURE.md
@@ -733,6 +733,7 @@ src/
 │   │   ├── tmux_lifecycle.rs
 │   │   ├── tmux_output_stream.rs
 │   │   ├── tmux_overload_retry.rs
+│   │   ├── tmux_placeholder_suppression.rs
 │   │   ├── tmux_reaper.rs
 │   │   ├── tmux_reattach_offsets.rs
 │   │   ├── tmux_restart_handoff.rs
